+++ ios/App/Components/Main.js
@@ -1,25 +1,25 @@
 import React, { Component } from 'react';
-import {
+import{
   StyleSheet,
   Text,
   TouchableHighlight,
   Image,
   View,
 } from 'react-native'
+import $ from "jquery";
 import api from './../Utils/api';
 import Icon from 'react-native-vector-icons/FontAwesome';
-const leftArrow = (<Icon name="arrow-left" size={30} color="#900" />)
-const rightArrow = (<Icon name="arrow-right" size={30} color="#900" />)
+const leftArrow = (<Icon name="arrow-left" size={30} color="#900" />);
+const rightArrow = (<Icon name="arrow-right" size={30} color="#900" />);
 
 const FBSDK = require('react-native-fbsdk');
 const {
   LoginButton,
+  GraphRequest,
+  GraphRequestManager,
   LoginManager,
   AccessToken
 } = FBSDK;
-
-var Digits = require('react-native-fabric-digits');
-var { DigitsLoginButton, DigitsLogoutButton } = Digits;
 
 var Login = React.createClass({
   // LoginButton.readPermissions = ["public_profile", "email", "user_friends"];
@@ -39,7 +39,18 @@ var Login = React.createClass({
               } else {
                 AccessToken.getCurrentAccessToken().then(
                   (data) => {
-                    alert(data.accessToken.toString())
+                    console.log(data);
+                    var userId = data.userID;
+                      
+                   const infoRequest = new GraphRequest(
+                    `/${userId}?fields=id,name,email,picture`,
+                    null,
+                    (err, res) => { err ? console.log(err) : console.log(res); },
+                    );  
+                     // TODO: in the callback here - instead of console.log(res) we'd send it to the DB
+                     // and pass it down via (props) so we keep reference to it.
+                    
+                  new GraphRequestManager().addRequest(infoRequest).start();
                   }
                 )
               }
@@ -75,19 +86,6 @@ var styles = StyleSheet.create({
     fontSize: 20,
     textAlign: 'center',
     margin: 10,
-  },
-  DigitsAuthenticateButton: {
-    height: 50,
-    width: 230,
-    backgroundColor: '#13988A',
-    justifyContent: 'center',
-    borderRadius: 5
-  },
-  DigitsAuthenticateButtonText: {
-    fontSize: 16,
-    color: '#fff',
-    alignSelf: 'center',
-    fontWeight: 'bold'
   }
 });
 
@@ -102,24 +100,7 @@ class Main extends Component {
       error: false,
       response: {}
     } // NOTE: this way we don't need 3 separate pages for images.
-  } // we can just swap in/out the images and text.
-    // which do we prefer? could use testing. Login button is the same on each.
-  
-  customHandleLogin() {
-    LoginManager.logInWithReadPermissions(['public_profile', 'email', 'user_friends']).then(
-      function(result) {
-        if (result.isCancelled) {
-          alert('Login cancelled');
-        } else {
-          alert('Login success with permissions: '
-            +result.grantedPermissions.toString());
-        }
-      },
-      function(error) {
-        alert('Login fail with error: ' + error);
-      }
-    );
-  }
+  } 
   
   nextImage(imageId) {
     if (this.state.imageId < 2) {
@@ -133,62 +114,7 @@ class Main extends Component {
     }
   }
 
-  completion(error, response) {
-    if (error && error.code !== 1) {
-      this.setState({ logged: false, error: true, response: {} });
-    } else if (response) {
-      var logged = JSON.stringify(response) === '{}' ? false : true;
-      this.setState({ logged: logged, error: false, response: response });
-    } else {
-      alert('Damn son another error on Digits');
-    }
-  }
-
   render() {
-    
-    var error = this.state.error ? <Text>An error occured.</Text> : null;
-    var content = this.state.logged ? 
-      (<View>
-        <Text>
-          Auth Token: {this.state.response.authToken}{'\n'}
-          Auth Token Secret: {this.state.response.authTokenSecret}{'\n\n'}
-        </Text>
-        <DigitsLogoutButton
-          completion={this.completion}
-          text="Logout"
-          buttonStyle={styles.DigitsAuthenticateButton}
-          textStyle={styles.DigitsAuthenticateButtonText}/>
-      </View>) : (<DigitsLoginButton
-        options={{
-          title: "Logging in is great",
-          phoneNumber: "+1",
-          appearance: {
-            backgroundColor: {
-              hex: "#ffffff",
-              alpha: 1.0
-            },
-            accentColor: {
-              hex: "#43a16f",
-              alpha: 0.7
-            },
-            headerFont: {
-              name: "Arial",
-              size: 16
-            },
-            labelFont: {
-              name: "Helvetica",
-              size: 18
-            },
-            bodyFont: {
-              name: "Helvetica",
-              size: 16
-            }
-          }
-        }}
-        completion={this.completion}
-        text="Login (using SMS)"
-        buttonStyle={styles.DigitsAuthenticateButton}
-        textStyle={styles.DigitsAuthenticateButtonText}/>);
     
     return (
       <View style={styles.container}>
@@ -205,13 +131,7 @@ class Main extends Component {
         </TouchableHighlight>
 
         <Login></Login>
-        
-        {error}
-        {content}
-        
-        <TouchableHighlight style={styles.container} onPress={this.customHandleLogin.bind(this)}>
-          <Text>click to log in</Text>
-        </TouchableHighlight>
+  
       </View>
     )
   }
