+++ package.json
@@ -6,11 +6,13 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
+    "jquery": "^3.1.0",
     "react": "15.2.1",
     "react-native": "0.29.1",
     "react-native-fabric": "^0.3.0",
     "react-native-fabric-digits": "github:jeanlebrument/react-native-fabric-digits",
     "react-native-fbsdk": "^0.2.2",
+    "react-native-swiper": "^1.4.6",
     "react-native-vector-icons": "^2.0.3"
   }
 }
